--- data/WiredTiger
@@ -1,2 +1,0 @@
-WiredTiger
-WiredTiger 10.0.2: (December 21, 2021)
