+++ server.js
@@ -1,64 +1,81 @@
-'use strict';
+import express from "express";
+import cors from "cors";
+import * as dotenv from "dotenv";
+import { NotFoundError, errorHandler } from "./middlewares/error-handler.js";
+import morgan from "morgan";
+import connectDb from "./config/db.js";
+import router from "./routes/routes.js";
+import path from 'path';
+import { fileURLToPath } from 'url';
+import { Server as socket } from "socket.io";
 
-var express = require('express'),
-app = express(),
-port = process.env.PORT || 3000,
-session = require('express-session'),
-bodyParser = require('body-parser'),
-jsonwebtoken = require("jsonwebtoken");
-const cloudinary = require('cloudinary').v2;
 
-cloudinary.config({ 
-    cloud_name: 'dfctbqlmj', 
-    api_key: '519927818185632', 
-    api_secret: 'UlPbXnJVOdxo_WgdW7MbB1GvKRA' 
+const app = express();
+
+const __filename = fileURLToPath(import.meta.url);
+
+const __dirname = path.dirname(__filename);
+
+const corsOptions = {
+  origin: 'http://localhost:3000',
+};
+dotenv.config();
+
+const hostname = process.env.SERVERURL;
+const port = process.env.SERVERPORT;
+
+//info on req : GET /route ms -25
+app.use(morgan("dev"));
+
+app.use(cors(corsOptions));
+connectDb();
+
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+app.use("/api", router);
+
+app.get("/api/*", function (req, res) {
+  res.sendFile(path.join(__dirname, "build/index.html"), function (err) {
+    if (err) {
+      res.status(500).send(err);
+      console.log(err);
+    }
+  });
+});
+
+app.use("/public/mp3", express.static(path.join(__dirname, "public/mp3")));
+app.use("/public/images", express.static(path.join(__dirname, "public/images")));
+
+const server = app.listen(port, () =>
+  console.log(`Server started on ${port}`)
+);
+
+const io = new socket(server, {
+  cors: {
+    origin: "http://localhost:3000",
+    credentials: true,
+  },
+});
+
+global.onlineUsers = new Map();
+io.on("connection", (socket) => {
+  global.chatSocket = socket;
+  socket.on("add-user", (userId) => {
+    onlineUsers.set(userId, socket.id);
   });
 
-
-const mongoose = require('mongoose');
-const option = {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-};
-
-app.use(session({
-    secret: 'secret', // used to sign the session ID cookie
-    resave: false, // don't save session if unmodified
-    saveUninitialized: false, // don't create session until something stored
-  }));
-
-
-mongoose.connect('mongodb+srv://Chawki32g:WP8uec9S5bPLjaP@c1.2qrjkua.mongodb.net/?retryWrites=true&w=majority', option).then(function(){
-    console.log("successfully connected to the database");
-}, function(err) {
-    console.log(err)
-});
-
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-
-app.use(function(req, res, next) {
-    if (req.headers && req.headers.cookie && req.headers.cookie.split('=')[0] === 'token') {
-        jsonwebtoken.verify(req.headers.cookie.split('=')[1], 'secret', function(err, decode) {
-            if (err) req.user = undefined;
-            req.user = decode;
-            next();
-        });
-    } else {
-        req.user = undefined;
-        next();
+  socket.on("send-msg", (data) => {
+    const sendUserSocket = onlineUsers.get(data.to);
+    if (sendUserSocket) {
+      socket.to(sendUserSocket).emit("msg-recieve", data.msg);
     }
+  });
 });
 
-app.use(require('./api/routes/routes'));
+app.use(NotFoundError);
+app.use(errorHandler);
 
-
-app.use(function(req, res) {
-  res.status(404).send({ url: req.originalUrl + ' not found' })
+app.listen(port, hostname, () => {
+  console.log(`Server running on ${hostname}:${port}`);
 });
-
-app.listen(port);
-
-console.log(' RESTful API server started on: ' + port);
-
-module.exports = app;
